ImageKitDemo upgrade to unified
Removed old “MonoMac” namespace qualifier from several types.  Changed
int to uint where required. Cast nint to int in call to AddImages.

+++ ImageKitDemoStep1/BrowseData.cs
@@ -2,9 +2,9 @@ using System;
 using System.IO;
 using System.Collections.Generic;
 using System.Linq;
-using MonoMac.Foundation;
-using MonoMac.AppKit;
-using MonoMac.ImageKit;
+using Foundation;
+using AppKit;
+using ImageKit;
 
 namespace ImageKitDemo
 {
@@ -38,13 +38,13 @@ namespace ImageKitDemo
 			//indexes will change, but I will put the items in after the item at destination
 			//FIXME - missing methods on NSIndexSet
 			//FIXME make an extension method on List<>
-			int destination = destinationIndex - indexes.Where (x => x < destinationIndex).Count ();
+			nint destination = destinationIndex - indexes.Where (x => x < (nuint)destinationIndex).Count ();
 			List<BrowseItem> movingImages = new List<BrowseItem> ();
 			foreach (int index in indexes)
 				movingImages.Add (images[index]);
 			foreach (BrowseItem item in movingImages)
 				images.Remove (item);
-			images.InsertRange (destination, movingImages);
+			images.InsertRange ((int)destination, movingImages);
 			aBrowser.ReloadData();
 			return true;
 		}
